+++ .env.example
@@ -10,3 +10,4 @@ LLM_PROVIDER=my_ai
 CACHE=false # true to enable llm response caching
 GENERATE_TESTS=false # true to generate tests
 GENERATE_METADATA=false # true to generate metadata in main pipeline
+PERPLEXITY_API_KEY=get from https://www.perplexity.ai/settings/api
