+++ prompts/3b-perplexity.md
@@ -1,0 +1,6 @@
+I need to implement the following functionality: 
+<requirement>
+{{prompt}}
+</requirement>
+
+Please provide detailed technical information, implementation approaches, best practices, and any relevant code examples or libraries that could help implement this.
