+++ src/local-index.ts
@@ -1,7 +1,7 @@
 import { parseArgs } from "jsr:@std/cli/parse-args";
 import "jsr:@std/dotenv/load";
+import { getOpenAIO4Mini } from "./Engines/index.ts";
 import { FileManager } from "./ShinkaiPipeline/FileManager.ts";
-import { getOpenAIO4Mini } from "./ShinkaiPipeline/llm-engines.ts";
 import { getTests } from "./ShinkaiPipeline/Requirement.ts";
 import { ShinkaiPipeline } from "./ShinkaiPipeline/ShinkaiPipeline.ts";
 import { Language } from "./ShinkaiPipeline/types.ts";
